@@ -15,68 +15,130 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 3.4.7+191
+version: 4.5.0+294
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  assets_audio_player:
-    git:
-      url: https://github.com/asmengistu/Flutter-AssetsAudioPlayer
-      ref: d5f2a5b928ae118154aaf3ac2f1df96e301cb4af
+  apivideo_live_stream: 1.0.7
+  assets_audio_player: 3.0.6
+  assets_audio_player_web: 3.0.6
   auto_size_text: 3.0.0
   badges: 2.0.2
-  built_collection: 5.1.1
-  built_value: 8.1.3
-  cached_network_image: 3.1.0+1
-  chewie: 1.2.2
-  cloud_firestore: 2.5.4
-  cloud_firestore_web: 2.6.8
-  cloud_functions: 3.1.1
-  dash_chat: 1.1.16
+  better_player: ^0.0.83
+  cached_network_image: 3.2.1
+  chewie: 1.4.0
+  cloud_firestore: 4.8.0
+  cloud_firestore_platform_interface: 5.15.0
+  cloud_firestore_web: 3.6.0
+  cloud_functions: 4.3.2
+  cloud_functions_platform_interface: 5.4.2
+  cloud_functions_web: 4.5.2
+  dash_chat: 
+    git:
+      url: https://github.com/FlutterFlow/flutterflow_dash_chat.git
+
+  device_info_plus: 8.0.0
+  device_info_plus_platform_interface: 7.0.0
+  dropdown_button2: 2.0.0
   easy_debounce: 2.0.1
-  equatable: 2.0.3
-  file_picker: 4.5.1
-  firebase_analytics: 9.1.0
-  firebase_auth: 3.2.0
-  firebase_core: 1.12.0
-  firebase_core_web: 1.5.4
-  firebase_messaging: 11.2.4
-  firebase_messaging_platform_interface: 3.3.1
-  firebase_messaging_web: 2.2.7
-  firebase_storage: 10.1.0
-  firebase_storage_web: 3.2.8
-  flutter_spinkit: 5.1.0
+  emoji_flag_converter: 1.1.0
+  equatable: 2.0.5
+  expandable: 5.0.1
+  file_picker: 5.2.6
+  firebase_analytics: 10.4.2
+  firebase_analytics_platform_interface: 3.6.2
+  firebase_analytics_web: 0.5.4+2
+  firebase_auth: 4.6.2
+  firebase_auth_platform_interface: 6.15.2
+  firebase_auth_web: 5.5.2
+  firebase_core: 2.13.1
+  firebase_core_platform_interface: 4.8.0
+  firebase_core_web: 2.5.0
+  firebase_crashlytics: 3.3.2
+  firebase_crashlytics_platform_interface: 3.6.2
+  firebase_messaging: 14.6.2
+  firebase_messaging_platform_interface: 4.5.2
+  firebase_messaging_web: 3.5.2
+  firebase_remote_config: 4.2.2
+  firebase_remote_config_platform_interface: 1.4.2
+  firebase_remote_config_web: 1.4.2
+  firebase_storage: 11.2.2
+  firebase_storage_platform_interface: 4.4.2
+  firebase_storage_web: 3.6.2
+  flutter_animate: 4.1.1+1
+  flutter_cache_manager: 3.3.0
+  flutter_plugin_android_lifecycle: 2.0.9
+  flutter_staggered_grid_view: 0.6.2
   font_awesome_flutter: 10.1.0
-  geolocator: 8.0.1
-  go_router: 3.1.0
-  google_fonts: 2.2.0
-  google_sign_in: 5.2.1
-  image_picker: 0.8.4+4
-  intl: 0.17.0
-  json_path: 0.3.1
-  json_serializable: 6.1.3
+  from_css_color: 2.0.0
+  geolocator: 9.0.1
+  geolocator_android: 4.1.0
+  geolocator_apple: 2.2.1
+  geolocator_platform_interface: 4.0.6
+  geolocator_web: 2.1.6
+  go_router: 7.1.1
+  google_fonts: 4.0.3
+  google_nav_bar: 5.0.5
+  google_sign_in: 6.0.2
+  google_sign_in_android: 6.1.8
+  google_sign_in_ios: 5.6.1
+  google_sign_in_platform_interface: 2.4.0
+  google_sign_in_web: 0.11.0+2
+  image_picker: 0.8.5+3
+  image_picker_android: 0.8.5+3
+  image_picker_for_web: 2.1.10
+  image_picker_ios: 0.8.6+1
+  image_picker_platform_interface: 2.6.2
+  internet_file: 1.1.0
+  intl: 0.18.0
+  json_path: 0.4.1
+  lottie: 1.2.1
   mime_type: 1.0.0
   page_transition: 2.0.4
-  percent_indicator: 4.0.0
-  photo_view: ^0.13.0
-  rxdart: 0.26.0
-  share_plus: 2.1.4
-  shared_preferences: 2.0.11
-  sign_in_with_apple:
-    git:
-      url: https://github.com/asmengistu/dart_packages
-      ref: ab18d53a80e0a4bb90f1ad84f37f1b408930f28d
-      path: packages/sign_in_with_apple/sign_in_with_apple/
-  table_calendar: 3.0.5
-  timeago: 3.1.0
-  url_launcher: 6.0.15
-  video_player: 2.2.7
+  path_provider: 2.0.14
+  path_provider_android: 2.0.25
+  path_provider_foundation: 2.2.2
+  path_provider_platform_interface: 2.0.6
+  pdfx: 2.3.0
+  percent_indicator: 4.2.2
+  photo_view: 0.14.0
+  plugin_platform_interface: 2.1.3
+  provider: 6.0.4
+  rxdart: 0.27.7
+  share_plus: 6.3.1
+  shared_preferences: 2.0.15
+  shared_preferences_android: 2.1.0
+  shared_preferences_ios: 2.1.1
+  shared_preferences_platform_interface: 2.2.0
+  shared_preferences_web: 2.1.0
+  sign_in_with_apple: 4.3.0
+  sign_in_with_apple_platform_interface: 1.0.0
+  sign_in_with_apple_web: 1.0.1
+  simple_gradient_text: 1.2.3
+  smooth_page_indicator: 1.0.1
+  sqflite: 2.2.6
+  stop_watch_timer: 1.5.0
+  stream_transform: 2.1.0
+  substring_highlight: 1.0.33
+  table_calendar: 3.0.9
+  text_search: 0.0.2
+  timeago: 3.2.2
+  url_launcher: 6.1.10
+  url_launcher_android: 6.0.27
+  url_launcher_ios: 6.1.4
+  url_launcher_platform_interface: 2.1.2
+  video_player: 2.6.0
+  video_player_android: 2.4.3
+  video_player_avfoundation: 2.4.2
+  video_player_platform_interface: 6.1.0
+  video_player_web: 2.0.15
+  wakelock: 0.6.2
 
 
   # The following adds the Cupertino Icons font to your application.
