@@ -1,9 +1,15 @@
 import 'dart:convert';
 import 'dart:io';
 import 'dart:core';
+import 'dart:typed_data';
 
+import 'package:collection/collection.dart';
 import 'package:http/http.dart' as http;
 import 'package:equatable/equatable.dart';
+import 'package:http_parser/http_parser.dart';
+import 'package:mime_type/mime_type.dart';
+
+import '../../flutter_flow/uploaded_file.dart';
 
 enum ApiCallType {
   GET,
@@ -18,6 +24,7 @@ enum BodyType {
   JSON,
   TEXT,
   X_WWW_FORM_URL_ENCODED,
+  MULTIPART,
 }
 
 class ApiCallRecord extends Equatable {
@@ -27,19 +34,60 @@ class ApiCallRecord extends Equatable {
   final String apiUrl;
   final Map<String, dynamic> headers;
   final Map<String, dynamic> params;
-  final String body;
-  final BodyType bodyType;
+  final String? body;
+  final BodyType? bodyType;
 
   @override
-  List<Object> get props => [callName, apiUrl, headers, params, body, bodyType];
+  List<Object?> get props =>
+      [callName, apiUrl, headers, params, body, bodyType];
 }
 
 class ApiCallResponse {
-  const ApiCallResponse(this.jsonBody, this.statusCode);
+  const ApiCallResponse(
+    this.jsonBody,
+    this.headers,
+    this.statusCode, {
+    this.response,
+  });
   final dynamic jsonBody;
+  final Map<String, String> headers;
   final int statusCode;
-  // Whether we recieved a 2xx status (which generally marks success).
+  final http.Response? response;
+  // Whether we received a 2xx status (which generally marks success).
   bool get succeeded => statusCode >= 200 && statusCode < 300;
+  String getHeader(String headerName) => headers[headerName] ?? '';
+  // Return the raw body from the response, or if this came from a cloud call
+  // and the body is not a string, then the json encoded body.
+  String get bodyText =>
+      response?.body ??
+      (jsonBody is String ? jsonBody as String : jsonEncode(jsonBody));
+
+  static ApiCallResponse fromHttpResponse(
+    http.Response response,
+    bool returnBody,
+    bool decodeUtf8,
+  ) {
+    var jsonBody;
+    try {
+      final responseBody = decodeUtf8 && returnBody
+          ? const Utf8Decoder().convert(response.bodyBytes)
+          : response.body;
+      jsonBody = returnBody ? json.decode(responseBody) : null;
+    } catch (_) {}
+    return ApiCallResponse(
+      jsonBody,
+      response.headers,
+      response.statusCode,
+      response: response,
+    );
+  }
+
+  static ApiCallResponse fromCloudCallResponse(Map<String, dynamic> response) =>
+      ApiCallResponse(
+        response['body'],
+        ApiManager.toStringMap(response['headers'] ?? {}),
+        response['statusCode'] ?? 400,
+      );
 }
 
 class ApiManager {
@@ -48,12 +96,12 @@ class ApiManager {
   // Cache that will ensure identical calls are not repeatedly made.
   static Map<ApiCallRecord, ApiCallResponse> _apiCache = {};
 
-  static ApiManager _instance;
+  static ApiManager? _instance;
   static ApiManager get instance => _instance ??= ApiManager._();
 
   // If your API calls need authentication, populate this field once
   // the user has authenticated. Alter this as needed.
-  static String _accessToken;
+  static String? _accessToken;
 
   // You may want to call this if, for example, you make a change to the
   // database and no longer want the cached result of a call that may
@@ -62,20 +110,11 @@ class ApiManager {
       .toSet()
       .forEach((k) => k.callName == callName ? _apiCache.remove(k) : null);
 
-  static Map<String, String> toStringMap(Map<String, dynamic> map) =>
-      map.map((key, value) => MapEntry(key, value.toString()));
+  static Map<String, String> toStringMap(Map map) =>
+      map.map((key, value) => MapEntry(key.toString(), value.toString()));
 
   static String asQueryParams(Map<String, dynamic> map) =>
       map.entries.map((e) => "${e.key}=${e.value}").join('&');
-
-  static ApiCallResponse createResponse(
-      http.Response response, bool returnBody) {
-    var jsonBody;
-    try {
-      jsonBody = returnBody ? json.decode(response.body) : null;
-    } catch (_) {}
-    return ApiCallResponse(jsonBody, response.statusCode);
-  }
 
   static Future<ApiCallResponse> urlRequest(
     ApiCallType callType,
@@ -83,6 +122,7 @@ class ApiManager {
     Map<String, dynamic> headers,
     Map<String, dynamic> params,
     bool returnBody,
+    bool decodeUtf8,
   ) async {
     if (params.isNotEmpty) {
       final lastUriPart = apiUrl.split('/').last;
@@ -93,7 +133,7 @@ class ApiManager {
     final makeRequest = callType == ApiCallType.GET ? http.get : http.delete;
     final response =
         await makeRequest(Uri.parse(apiUrl), headers: toStringMap(headers));
-    return createResponse(response, returnBody);
+    return ApiCallResponse.fromHttpResponse(response, returnBody, decodeUtf8);
   }
 
   static Future<ApiCallResponse> requestWithBody(
@@ -101,36 +141,102 @@ class ApiManager {
     String apiUrl,
     Map<String, dynamic> headers,
     Map<String, dynamic> params,
-    String body,
-    BodyType bodyType,
+    String? body,
+    BodyType? bodyType,
     bool returnBody,
+    bool encodeBodyUtf8,
+    bool decodeUtf8,
   ) async {
     assert(
       {ApiCallType.POST, ApiCallType.PUT, ApiCallType.PATCH}.contains(type),
       'Invalid ApiCallType $type for request with body',
     );
-    final postBody = createBody(headers, params, body, bodyType);
+    final postBody =
+        createBody(headers, params, body, bodyType, encodeBodyUtf8);
+
+    if (bodyType == BodyType.MULTIPART) {
+      return multipartRequest(
+          type, apiUrl, headers, params, returnBody, decodeUtf8);
+    }
+
     final requestFn = {
       ApiCallType.POST: http.post,
       ApiCallType.PUT: http.put,
       ApiCallType.PATCH: http.patch,
-    }[type];
+    }[type]!;
     final response = await requestFn(Uri.parse(apiUrl),
         headers: toStringMap(headers), body: postBody);
-    return createResponse(response, returnBody);
+    return ApiCallResponse.fromHttpResponse(response, returnBody, decodeUtf8);
+  }
+
+  static Future<ApiCallResponse> multipartRequest(
+    ApiCallType? type,
+    String apiUrl,
+    Map<String, dynamic> headers,
+    Map<String, dynamic> params,
+    bool returnBody,
+    bool decodeUtf8,
+  ) async {
+    assert(
+      {ApiCallType.POST, ApiCallType.PUT, ApiCallType.PATCH}.contains(type),
+      'Invalid ApiCallType $type for request with body',
+    );
+    bool Function(dynamic) _isFile = (e) =>
+        e is FFUploadedFile ||
+        e is List<FFUploadedFile> ||
+        (e is List && e.firstOrNull is FFUploadedFile);
+
+    final nonFileParams = toStringMap(
+        Map.fromEntries(params.entries.where((e) => !_isFile(e.value))));
+
+    List<http.MultipartFile> files = [];
+    params.entries.where((e) => _isFile(e.value)).forEach((e) {
+      final param = e.value;
+      final uploadedFiles = param is List
+          ? param as List<FFUploadedFile>
+          : [param as FFUploadedFile];
+      uploadedFiles.forEach((uploadedFile) => files.add(
+            http.MultipartFile.fromBytes(
+              e.key,
+              uploadedFile.bytes ?? Uint8List.fromList([]),
+              filename: uploadedFile.name,
+              contentType: _getMediaType(uploadedFile.name),
+            ),
+          ));
+    });
+
+    final request = http.MultipartRequest(
+        type.toString().split('.').last, Uri.parse(apiUrl))
+      ..headers.addAll(toStringMap(headers))
+      ..files.addAll(files);
+    nonFileParams.forEach((key, value) => request.fields[key] = value);
+
+    final response = await http.Response.fromStream(await request.send());
+    return ApiCallResponse.fromHttpResponse(response, returnBody, decodeUtf8);
+  }
+
+  static MediaType? _getMediaType(String? filename) {
+    final contentType = mime(filename);
+    if (contentType == null) {
+      return null;
+    }
+    final parts = contentType.split('/');
+    if (parts.length != 2) {
+      return null;
+    }
+    return MediaType(parts.first, parts.last);
   }
 
   static dynamic createBody(
     Map<String, dynamic> headers,
-    Map<String, dynamic> params,
-    String body,
-    BodyType bodyType,
+    Map<String, dynamic>? params,
+    String? body,
+    BodyType? bodyType,
+    bool encodeBodyUtf8,
   ) {
-    String contentType;
+    String? contentType;
     dynamic postBody;
     switch (bodyType) {
-      case BodyType.NONE:
-        break;
       case BodyType.JSON:
         contentType = 'application/json';
         postBody = body ?? json.encode(params ?? {});
@@ -141,23 +247,37 @@ class ApiManager {
         break;
       case BodyType.X_WWW_FORM_URL_ENCODED:
         contentType = 'application/x-www-form-urlencoded';
-        postBody = toStringMap(params);
+        postBody = toStringMap(params ?? {});
+        break;
+      case BodyType.MULTIPART:
+        contentType = 'multipart/form-data';
+        postBody = params;
+        break;
+      case BodyType.NONE:
+      case null:
+        break;
     }
-    if (contentType != null) {
+    // Set "Content-Type" header if it was previously unset.
+    if (contentType != null &&
+        !headers.keys.any((h) => h.toLowerCase() == 'content-type')) {
       headers['Content-Type'] = contentType;
     }
-    return postBody;
+    return encodeBodyUtf8 && postBody is String
+        ? utf8.encode(postBody)
+        : postBody;
   }
 
   Future<ApiCallResponse> makeApiCall({
-    String callName,
-    String apiUrl,
-    ApiCallType callType,
+    required String callName,
+    required String apiUrl,
+    required ApiCallType callType,
     Map<String, dynamic> headers = const {},
     Map<String, dynamic> params = const {},
-    String body,
-    BodyType bodyType,
-    bool returnBody,
+    String? body,
+    BodyType? bodyType,
+    bool returnBody = true,
+    bool encodeBodyUtf8 = false,
+    bool decodeUtf8 = false,
     bool cache = false,
   }) async {
     final callRecord =
@@ -173,26 +293,41 @@ class ApiManager {
     // If we've already made this exact call before and caching is on,
     // return the cached result.
     if (cache && _apiCache.containsKey(callRecord)) {
-      return _apiCache[callRecord];
+      return _apiCache[callRecord]!;
     }
 
     ApiCallResponse result;
     switch (callType) {
       case ApiCallType.GET:
       case ApiCallType.DELETE:
-        result =
-            await urlRequest(callType, apiUrl, headers, params, returnBody);
+        result = await urlRequest(
+          callType,
+          apiUrl,
+          headers,
+          params,
+          returnBody,
+          decodeUtf8,
+        );
         break;
       case ApiCallType.POST:
       case ApiCallType.PUT:
       case ApiCallType.PATCH:
         result = await requestWithBody(
-            callType, apiUrl, headers, params, body, bodyType, returnBody);
+          callType,
+          apiUrl,
+          headers,
+          params,
+          body,
+          bodyType,
+          returnBody,
+          encodeBodyUtf8,
+          decodeUtf8,
+        );
         break;
     }
 
     // If caching is on, cache the result (if present).
-    if (cache && result != null) {
+    if (cache) {
       _apiCache[callRecord] = result;
     }
 
