@@ -1,150 +1,93 @@
-import 'dart:math';
-import 'dart:ui';
-
-import 'package:flutter/animation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_animate/flutter_animate.dart';
 
 enum AnimationTrigger {
   onPageLoad,
   onActionTrigger,
 }
 
-class AnimationState {
-  AnimationState({
-    this.offset = const Offset(0, 0),
-    this.opacity = 1,
-    this.scale = 1,
-  });
-  final Offset offset;
-  final double opacity;
-  final double scale;
-}
-
 class AnimationInfo {
   AnimationInfo({
-    this.curve = Curves.easeInOut,
-    @required this.trigger,
-    @required this.duration,
-    this.delay = 0,
-    this.hideBeforeAnimating = true,
-    this.fadeIn = false,
-    this.initialState,
-    this.finalState,
+    required this.trigger,
+    required this.effects,
+    this.loop = false,
+    this.reverse = false,
+    this.applyInitialState = true,
   });
-
-  final Curve curve;
   final AnimationTrigger trigger;
-  final int duration;
-  final int delay;
-  final bool hideBeforeAnimating;
-  final bool fadeIn;
-  final AnimationState initialState;
-  final AnimationState finalState;
-  CurvedAnimation curvedAnimation;
+  final List<Effect<dynamic>> effects;
+  final bool applyInitialState;
+  final bool loop;
+  final bool reverse;
+  late AnimationController controller;
 }
 
 void createAnimation(AnimationInfo animation, TickerProvider vsync) {
-  animation.curvedAnimation = CurvedAnimation(
-    parent: AnimationController(
-      duration: Duration(milliseconds: animation.duration),
-      vsync: vsync,
-    ),
-    curve: animation.curve,
-  );
+  final newController = AnimationController(vsync: vsync);
+  animation.controller = newController;
 }
 
-void startPageLoadAnimations(
-    Iterable<AnimationInfo> animations, TickerProvider vsync) {
-  animations.forEach((animation) async {
-    createAnimation(animation, vsync);
-    await Future.delayed(
-      Duration(milliseconds: animation.delay),
-      () => (animation.curvedAnimation.parent as AnimationController)
-          .forward(from: 0.0),
-    );
-  });
-}
-
-void setupTriggerAnimations(
-    Iterable<AnimationInfo> animations, TickerProvider vsync) {
-  animations.forEach((animation) {
-    createAnimation(animation, vsync);
-  });
+void setupAnimations(Iterable<AnimationInfo> animations, TickerProvider vsync) {
+  animations.forEach((animation) => createAnimation(animation, vsync));
 }
 
 extension AnimatedWidgetExtension on Widget {
-  Widget animated(Iterable<AnimationInfo> animationInfos) {
-    final animationInfo = animationInfos.first;
-    return AnimatedBuilder(
-      animation: animationInfo.curvedAnimation,
-      builder: (context, child) {
-        if (child == null) {
-          return Container();
-        }
-        // On Action Trigger animations are in this state when
-        // they are first loaded, but before they are triggered.
-        // The widget should remain as it is.
-        if (animationInfo.curvedAnimation.status == AnimationStatus.dismissed) {
-          if (animationInfo.hideBeforeAnimating) {
-            return Opacity(
-              opacity: 0,
-              child: child,
-            );
-          } else {
-            return child;
-          }
-        }
-        var returnedWidget = child;
-        if (animationInfo.initialState.offset.dx != 0 ||
-            animationInfo.initialState.offset.dy != 0 ||
-            animationInfo.finalState.offset.dx != 0 ||
-            animationInfo.finalState.offset.dy != 0) {
-          final xRange = animationInfo.finalState.offset.dx -
-              animationInfo.initialState.offset.dx;
-          final yRange = animationInfo.finalState.offset.dy -
-              animationInfo.initialState.offset.dy;
-          final xDelta = xRange * animationInfo.curvedAnimation.value;
-          final yDelta = yRange * animationInfo.curvedAnimation.value;
-          returnedWidget = Transform.translate(
-            offset: Offset(
-              animationInfo.initialState.offset.dx + xDelta,
-              animationInfo.initialState.offset.dy + yDelta,
-            ),
-            child: returnedWidget,
-          );
-        }
-        if (animationInfo.initialState.scale != 1 ||
-            animationInfo.finalState.scale != 1) {
-          final range =
-              animationInfo.finalState.scale - animationInfo.initialState.scale;
-          final delta = range * animationInfo.curvedAnimation.value;
-          final scale = animationInfo.initialState.scale + delta;
+  Widget animateOnPageLoad(AnimationInfo animationInfo) => Animate(
+      effects: animationInfo.effects,
+      child: this,
+      onPlay: (controller) => animationInfo.loop
+          ? controller.repeat(reverse: animationInfo.reverse)
+          : null,
+      onComplete: (controller) => !animationInfo.loop && animationInfo.reverse
+          ? controller.reverse()
+          : null);
 
-          returnedWidget = Transform.scale(
-            scale: scale,
-            child: returnedWidget,
-          );
-        }
-        if (animationInfo.fadeIn) {
-          final opacityRange = animationInfo.finalState.opacity -
-              animationInfo.initialState.opacity;
-          final opacityDelta =
-              animationInfo.curvedAnimation.value * opacityRange;
-          final opacity = animationInfo.initialState.opacity + opacityDelta;
+  Widget animateOnActionTrigger(
+    AnimationInfo animationInfo, {
+    bool hasBeenTriggered = false,
+  }) =>
+      hasBeenTriggered || animationInfo.applyInitialState
+          ? Animate(
+              controller: animationInfo.controller,
+              autoPlay: false,
+              effects: animationInfo.effects,
+              child: this)
+          : this;
+}
 
-          returnedWidget = Opacity(
-            // In cases where the child tree has a Material widget with elevation,
-            // opacity animations may result in sudden box shadow "glitches"
-            // To prevent this, opacity is animated up to but NOT including 1.0.
-            // It is impossible to tell the difference between 0.998 and 1.0 opacity.
-            opacity: min(0.998, opacity),
-            child: returnedWidget,
-          );
-        }
-        return returnedWidget;
-      },
-      child:
-          animationInfos.length > 1 ? animated(animationInfos.skip(1)) : this,
+class TiltEffect extends Effect<Offset> {
+  const TiltEffect({
+    Duration? delay,
+    Duration? duration,
+    Curve? curve,
+    Offset? begin,
+    Offset? end,
+  }) : super(
+          delay: delay,
+          duration: duration,
+          curve: curve,
+          begin: begin ?? const Offset(0.0, 0.0),
+          end: end ?? const Offset(0.0, 0.0),
+        );
+
+  @override
+  Widget build(
+    BuildContext context,
+    Widget child,
+    AnimationController controller,
+    EffectEntry entry,
+  ) {
+    Animation<Offset> animation = buildAnimation(controller, entry);
+    return getOptimizedBuilder<Offset>(
+      animation: animation,
+      builder: (_, __) => Transform(
+        transform: Matrix4.identity()
+          ..setEntry(3, 2, 0.001)
+          ..rotateX(animation.value.dx)
+          ..rotateY(animation.value.dy),
+        alignment: Alignment.center,
+        child: child,
+      ),
     );
   }
 }
