@@ -20,17 +20,21 @@ import 'dart:math' as math;
 import 'package:assets_audio_player/assets_audio_player.dart';
 import 'package:flutter/material.dart';
 
+import '/flutter_flow/flutter_flow_util.dart' show routeObserver;
+
 export 'package:assets_audio_player/assets_audio_player.dart';
 
 class FlutterFlowAudioPlayer extends StatefulWidget {
   const FlutterFlowAudioPlayer({
-    @required this.audio,
-    this.titleTextStyle,
-    this.playbackDurationTextStyle,
-    this.fillColor,
-    this.playbackButtonColor,
-    this.activeTrackColor,
-    this.elevation,
+    required this.audio,
+    required this.titleTextStyle,
+    required this.playbackDurationTextStyle,
+    required this.fillColor,
+    required this.playbackButtonColor,
+    required this.activeTrackColor,
+    required this.elevation,
+    this.pauseOnNavigate = true,
+    required this.playInBackground,
   });
 
   final Audio audio;
@@ -40,13 +44,17 @@ class FlutterFlowAudioPlayer extends StatefulWidget {
   final Color playbackButtonColor;
   final Color activeTrackColor;
   final double elevation;
+  final bool pauseOnNavigate;
+  final PlayInBackground playInBackground;
 
   @override
   _FlutterFlowAudioPlayerState createState() => _FlutterFlowAudioPlayerState();
 }
 
-class _FlutterFlowAudioPlayerState extends State<FlutterFlowAudioPlayer> {
-  AssetsAudioPlayer _assetsAudioPlayer;
+class _FlutterFlowAudioPlayerState extends State<FlutterFlowAudioPlayer>
+    with RouteAware {
+  AssetsAudioPlayer? _assetsAudioPlayer;
+  bool _subscribedRoute = false;
 
   @override
   void initState() {
@@ -57,21 +65,51 @@ class _FlutterFlowAudioPlayerState extends State<FlutterFlowAudioPlayer> {
   Future openPlayer() async {
     _assetsAudioPlayer ??=
         AssetsAudioPlayer.withId(generateRandomAlphaNumericString());
-    if (_assetsAudioPlayer.playlist != null) {
-      _assetsAudioPlayer.playlist.replaceAt(0, (oldAudio) => widget.audio);
+    if (_assetsAudioPlayer?.playlist != null) {
+      _assetsAudioPlayer!.playlist!.replaceAt(0, (oldAudio) => widget.audio);
     } else {
-      await _assetsAudioPlayer.open(
+      await _assetsAudioPlayer!.open(
         Playlist(audios: [widget.audio]),
         autoStart: false,
-        playInBackground: PlayInBackground.disabledRestoreOnForeground,
+        playInBackground: widget.playInBackground,
       );
     }
   }
 
   @override
   void dispose() {
+    if (_subscribedRoute) {
+      routeObserver.unsubscribe(this);
+    }
     _assetsAudioPlayer?.dispose();
     super.dispose();
+  }
+
+  @override
+  void didUpdateWidget(FlutterFlowAudioPlayer old) {
+    super.didUpdateWidget(old);
+    final changed = old.audio.path != widget.audio.path ||
+        old.audio.audioType != widget.audio.audioType;
+    final isPlaying = _assetsAudioPlayer?.isPlaying.value ?? false;
+    if (changed && !isPlaying) {
+      openPlayer();
+    }
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    if (widget.pauseOnNavigate && ModalRoute.of(context) is PageRoute) {
+      _subscribedRoute = true;
+      routeObserver.subscribe(this, ModalRoute.of(context)!);
+    }
+  }
+
+  @override
+  void didPushNext() {
+    if (widget.pauseOnNavigate) {
+      _assetsAudioPlayer?.pause();
+    }
   }
 
   Duration currentPosition(RealtimePlayingInfos infos) => infos.currentPosition;
@@ -85,9 +123,9 @@ class _FlutterFlowAudioPlayerState extends State<FlutterFlowAudioPlayer> {
 
   @override
   Widget build(BuildContext context) =>
-      _assetsAudioPlayer.builderRealtimePlayingInfos(
+      _assetsAudioPlayer!.builderRealtimePlayingInfos(
           builder: (context, infos) => PlayerBuilder.isPlaying(
-              player: _assetsAudioPlayer,
+              player: _assetsAudioPlayer!,
               builder: (context, isPlaying) {
                 final childWidget = Container(
                   width: double.infinity,
@@ -125,7 +163,7 @@ class _FlutterFlowAudioPlayerState extends State<FlutterFlowAudioPlayer> {
                             child: Material(
                               color: Colors.transparent,
                               child: IconButton(
-                                onPressed: _assetsAudioPlayer.playOrPause,
+                                onPressed: _assetsAudioPlayer!.playOrPause,
                                 icon: Icon(
                                   (isPlaying)
                                       ? Icons.pause_circle_filled_rounded
@@ -143,7 +181,7 @@ class _FlutterFlowAudioPlayerState extends State<FlutterFlowAudioPlayer> {
                         currentPosition: currentPosition(infos),
                         duration: duration(infos),
                         seekTo: (to) {
-                          _assetsAudioPlayer.seek(to);
+                          _assetsAudioPlayer!.seek(to);
                         },
                         activeTrackColor: widget.activeTrackColor,
                       ),
@@ -162,10 +200,10 @@ class _FlutterFlowAudioPlayerState extends State<FlutterFlowAudioPlayer> {
 
 class PositionSeekWidget extends StatefulWidget {
   const PositionSeekWidget({
-    @required this.currentPosition,
-    @required this.duration,
-    @required this.seekTo,
-    @required this.activeTrackColor,
+    required this.currentPosition,
+    required this.duration,
+    required this.seekTo,
+    required this.activeTrackColor,
   });
 
   final Duration currentPosition;
@@ -178,7 +216,7 @@ class PositionSeekWidget extends StatefulWidget {
 }
 
 class _PositionSeekWidgetState extends State<PositionSeekWidget> {
-  Duration _visibleValue;
+  late Duration _visibleValue;
   bool listenOnlyUserInteraction = false;
   double get percent => widget.duration.inMilliseconds == 0
       ? 0
@@ -245,11 +283,12 @@ class FlutterFlowRoundedRectSliderTrackShape extends SliderTrackShape
   void paint(
     PaintingContext context,
     Offset offset, {
-    @required RenderBox parentBox,
-    @required SliderThemeData sliderTheme,
-    @required Animation<double> enableAnimation,
-    @required TextDirection textDirection,
-    @required Offset thumbCenter,
+    required RenderBox parentBox,
+    Offset? secondaryOffset,
+    required SliderThemeData sliderTheme,
+    required Animation<double> enableAnimation,
+    required TextDirection textDirection,
+    required Offset thumbCenter,
     bool isDiscrete = false,
     bool isEnabled = false,
     double additionalActiveTrackHeight = 0,
@@ -269,7 +308,7 @@ class FlutterFlowRoundedRectSliderTrackShape extends SliderTrackShape
     // If the slider [SliderThemeData.trackHeight] is less than or equal to 0,
     // then it makes no difference whether the track is painted or not,
     // therefore the painting  can be a no-op.
-    if (sliderTheme.trackHeight == null || sliderTheme.trackHeight <= 0) {
+    if (sliderTheme.trackHeight == null || sliderTheme.trackHeight! <= 0) {
       return;
     }
 
@@ -282,9 +321,9 @@ class FlutterFlowRoundedRectSliderTrackShape extends SliderTrackShape
         begin: sliderTheme.disabledInactiveTrackColor,
         end: sliderTheme.inactiveTrackColor);
     final Paint activePaint = Paint()
-      ..color = activeTrackColorTween.evaluate(enableAnimation);
+      ..color = activeTrackColorTween.evaluate(enableAnimation)!;
     final Paint inactivePaint = Paint()
-      ..color = inactiveTrackColorTween.evaluate(enableAnimation);
+      ..color = inactiveTrackColorTween.evaluate(enableAnimation)!;
     final Paint leftTrackPaint = activePaint;
     final Paint rightTrackPaint = inactivePaint;
 
