@@ -1,49 +1,161 @@
 import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
-class FlutterFlowIconButton extends StatelessWidget {
-  const FlutterFlowIconButton(
-      {Key key,
-      this.borderColor,
-      this.borderRadius,
-      this.borderWidth,
-      this.buttonSize,
-      this.fillColor,
-      this.icon,
-      this.onPressed})
-      : super(key: key);
+class FlutterFlowIconButton extends StatefulWidget {
+  const FlutterFlowIconButton({
+    Key? key,
+    required this.icon,
+    this.borderColor,
+    this.borderRadius,
+    this.borderWidth,
+    this.buttonSize,
+    this.fillColor,
+    this.disabledColor,
+    this.disabledIconColor,
+    this.hoverColor,
+    this.hoverIconColor,
+    this.onPressed,
+    this.showLoadingIndicator = false,
+  }) : super(key: key);
 
-  final double borderRadius;
-  final double buttonSize;
-  final Color fillColor;
-  final Color borderColor;
-  final double borderWidth;
   final Widget icon;
-  final void Function() onPressed;
+  final double? borderRadius;
+  final double? buttonSize;
+  final Color? fillColor;
+  final Color? disabledColor;
+  final Color? disabledIconColor;
+  final Color? hoverColor;
+  final Color? hoverIconColor;
+  final Color? borderColor;
+  final double? borderWidth;
+  final bool showLoadingIndicator;
+  final Function()? onPressed;
 
   @override
-  Widget build(BuildContext context) => Material(
-        borderRadius:
-            borderRadius != null ? BorderRadius.circular(borderRadius) : null,
-        color: Colors.transparent,
-        clipBehavior: Clip.antiAlias,
-        child: Ink(
-          width: buttonSize,
-          height: buttonSize,
-          decoration: BoxDecoration(
-            color: fillColor,
-            border: Border.all(
-              color: borderColor ?? Colors.transparent,
-              width: borderWidth ?? 0,
+  State<FlutterFlowIconButton> createState() => _FlutterFlowIconButtonState();
+}
+
+class _FlutterFlowIconButtonState extends State<FlutterFlowIconButton> {
+  bool loading = false;
+  late double? iconSize;
+  late Color? iconColor;
+  late Widget effectiveIcon;
+
+  @override
+  void initState() {
+    super.initState();
+    _updateIcon();
+  }
+
+  @override
+  void didUpdateWidget(FlutterFlowIconButton oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    _updateIcon();
+  }
+
+  void _updateIcon() {
+    final isFontAwesome = widget.icon is FaIcon;
+    if (isFontAwesome) {
+      FaIcon icon = widget.icon as FaIcon;
+      effectiveIcon = FaIcon(
+        icon.icon,
+        size: icon.size,
+      );
+      iconSize = icon.size;
+      iconColor = icon.color;
+    } else {
+      Icon icon = widget.icon as Icon;
+      effectiveIcon = Icon(
+        icon.icon,
+        size: icon.size,
+      );
+      iconSize = icon.size;
+      iconColor = icon.color;
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    ButtonStyle style = ButtonStyle(
+      shape: MaterialStateProperty.resolveWith<OutlinedBorder>(
+        (states) {
+          return RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(widget.borderRadius ?? 0),
+            side: BorderSide(
+              color: widget.borderColor ?? Colors.transparent,
+              width: widget.borderWidth ?? 0,
             ),
-            borderRadius: borderRadius != null
-                ? BorderRadius.circular(borderRadius)
-                : null,
-          ),
+          );
+        },
+      ),
+      iconColor: MaterialStateProperty.resolveWith<Color?>(
+        (states) {
+          if (states.contains(MaterialState.disabled) &&
+              widget.disabledIconColor != null) {
+            return widget.disabledIconColor;
+          }
+          if (states.contains(MaterialState.hovered) &&
+              widget.hoverIconColor != null) {
+            return widget.hoverIconColor;
+          }
+          return iconColor;
+        },
+      ),
+      backgroundColor: MaterialStateProperty.resolveWith<Color?>(
+        (states) {
+          if (states.contains(MaterialState.disabled) &&
+              widget.disabledColor != null) {
+            return widget.disabledColor;
+          }
+          if (states.contains(MaterialState.hovered) &&
+              widget.hoverColor != null) {
+            return widget.hoverColor;
+          }
+
+          return widget.fillColor;
+        },
+      ),
+    );
+
+    return SizedBox(
+      width: widget.buttonSize,
+      height: widget.buttonSize,
+      child: Theme(
+        data: Theme.of(context).copyWith(useMaterial3: true),
+        child: IgnorePointer(
+          ignoring: (widget.showLoadingIndicator && loading),
           child: IconButton(
-            icon: icon,
-            onPressed: onPressed,
-            splashRadius: buttonSize,
+            icon: (widget.showLoadingIndicator && loading)
+                ? Container(
+                    width: iconSize,
+                    height: iconSize,
+                    child: CircularProgressIndicator(
+                      valueColor: AlwaysStoppedAnimation<Color>(
+                        iconColor ?? Colors.white,
+                      ),
+                    ),
+                  )
+                : effectiveIcon,
+            onPressed: widget.onPressed == null
+                ? null
+                : () async {
+                    if (loading) {
+                      return;
+                    }
+                    setState(() => loading = true);
+                    try {
+                      await widget.onPressed!();
+                    } finally {
+                      if (mounted) {
+                        setState(() => loading = false);
+                      }
+                    }
+                  },
+            splashRadius: widget.buttonSize,
+            style: style,
           ),
         ),
-      );
+      ),
+    );
+  }
 }
