@@ -6,7 +6,7 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 const kThemeModeKey = '__theme_mode__';
-SharedPreferences _prefs;
+SharedPreferences? _prefs;
 
 abstract class FlutterFlowTheme {
   static Future initialize() async =>
@@ -24,135 +24,387 @@ abstract class FlutterFlowTheme {
       ? _prefs?.remove(kThemeModeKey)
       : _prefs?.setBool(kThemeModeKey, mode == ThemeMode.dark);
 
-  static FlutterFlowTheme of(BuildContext context) =>
-      Theme.of(context).brightness == Brightness.dark
-          ? DarkModeTheme()
-          : LightModeTheme();
+  static FlutterFlowTheme of(BuildContext context) {
+    return Theme.of(context).brightness == Brightness.dark
+        ? DarkModeTheme()
+        : LightModeTheme();
+  }
 
-  Color primaryColor;
-  Color secondaryColor;
-  Color tertiaryColor;
-  Color alternate;
-  Color primaryBackground;
-  Color secondaryBackground;
-  Color primaryText;
-  Color secondaryText;
+  @Deprecated('Use primary instead')
+  Color get primaryColor => primary;
+  @Deprecated('Use secondary instead')
+  Color get secondaryColor => secondary;
+  @Deprecated('Use tertiary instead')
+  Color get tertiaryColor => tertiary;
 
-  Color primaryBtnText;
-  Color lineColor;
-  Color grayIcon;
-  Color gray200;
-  Color gray600;
-  Color black600;
-  Color tertiary400;
-  Color textColor;
-  Color backgroundComponents;
-  Color headerGray;
+  late Color primary;
+  late Color secondary;
+  late Color tertiary;
+  late Color alternate;
+  late Color primaryText;
+  late Color secondaryText;
+  late Color primaryBackground;
+  late Color secondaryBackground;
+  late Color accent1;
+  late Color accent2;
+  late Color accent3;
+  late Color accent4;
+  late Color success;
+  late Color warning;
+  late Color error;
+  late Color info;
 
-  TextStyle get title1 => GoogleFonts.getFont(
-        'Montserrat',
-        color: primaryText,
-        fontWeight: FontWeight.w600,
-        fontSize: 24,
-      );
-  TextStyle get title2 => GoogleFonts.getFont(
-        'Montserrat',
-        color: secondaryText,
-        fontWeight: FontWeight.w600,
-        fontSize: 22,
-      );
-  TextStyle get title3 => GoogleFonts.getFont(
-        'Montserrat',
-        color: primaryText,
-        fontWeight: FontWeight.w600,
-        fontSize: 20,
-      );
-  TextStyle get subtitle1 => GoogleFonts.getFont(
-        'Montserrat',
-        color: primaryText,
-        fontWeight: FontWeight.w600,
-        fontSize: 18,
-      );
-  TextStyle get subtitle2 => GoogleFonts.getFont(
-        'Montserrat',
-        color: secondaryText,
-        fontWeight: FontWeight.w600,
-        fontSize: 16,
-      );
-  TextStyle get bodyText1 => GoogleFonts.getFont(
-        'Montserrat',
-        color: primaryText,
-        fontWeight: FontWeight.w600,
-        fontSize: 14,
-      );
-  TextStyle get bodyText2 => GoogleFonts.getFont(
-        'Montserrat',
-        color: secondaryText,
-        fontWeight: FontWeight.w600,
-        fontSize: 14,
-      );
+  late Color primaryBtnText;
+  late Color lineColor;
+  late Color grayIcon;
+  late Color gray200;
+  late Color gray600;
+  late Color black600;
+  late Color tertiary400;
+  late Color textColor;
+  late Color backgroundComponents;
+  late Color headerGray;
+  late Color btnText;
+  late Color customColor3;
+  late Color customColor4;
+  late Color white;
+  late Color background;
+  late Color customColor1;
+  late Color overlay;
+  late Color primary600;
+  late Color secondary600;
+  late Color tertiary600;
+  late Color darkBGstatic;
+  late Color secondary30;
+  late Color overlay0;
+  late Color primary30;
+
+  @Deprecated('Use displaySmallFamily instead')
+  String get title1Family => displaySmallFamily;
+  @Deprecated('Use displaySmall instead')
+  TextStyle get title1 => typography.displaySmall;
+  @Deprecated('Use headlineMediumFamily instead')
+  String get title2Family => typography.headlineMediumFamily;
+  @Deprecated('Use headlineMedium instead')
+  TextStyle get title2 => typography.headlineMedium;
+  @Deprecated('Use headlineSmallFamily instead')
+  String get title3Family => typography.headlineSmallFamily;
+  @Deprecated('Use headlineSmall instead')
+  TextStyle get title3 => typography.headlineSmall;
+  @Deprecated('Use titleMediumFamily instead')
+  String get subtitle1Family => typography.titleMediumFamily;
+  @Deprecated('Use titleMedium instead')
+  TextStyle get subtitle1 => typography.titleMedium;
+  @Deprecated('Use titleSmallFamily instead')
+  String get subtitle2Family => typography.titleSmallFamily;
+  @Deprecated('Use titleSmall instead')
+  TextStyle get subtitle2 => typography.titleSmall;
+  @Deprecated('Use bodyMediumFamily instead')
+  String get bodyText1Family => typography.bodyMediumFamily;
+  @Deprecated('Use bodyMedium instead')
+  TextStyle get bodyText1 => typography.bodyMedium;
+  @Deprecated('Use bodySmallFamily instead')
+  String get bodyText2Family => typography.bodySmallFamily;
+  @Deprecated('Use bodySmall instead')
+  TextStyle get bodyText2 => typography.bodySmall;
+
+  String get displayLargeFamily => typography.displayLargeFamily;
+  TextStyle get displayLarge => typography.displayLarge;
+  String get displayMediumFamily => typography.displayMediumFamily;
+  TextStyle get displayMedium => typography.displayMedium;
+  String get displaySmallFamily => typography.displaySmallFamily;
+  TextStyle get displaySmall => typography.displaySmall;
+  String get headlineLargeFamily => typography.headlineLargeFamily;
+  TextStyle get headlineLarge => typography.headlineLarge;
+  String get headlineMediumFamily => typography.headlineMediumFamily;
+  TextStyle get headlineMedium => typography.headlineMedium;
+  String get headlineSmallFamily => typography.headlineSmallFamily;
+  TextStyle get headlineSmall => typography.headlineSmall;
+  String get titleLargeFamily => typography.titleLargeFamily;
+  TextStyle get titleLarge => typography.titleLarge;
+  String get titleMediumFamily => typography.titleMediumFamily;
+  TextStyle get titleMedium => typography.titleMedium;
+  String get titleSmallFamily => typography.titleSmallFamily;
+  TextStyle get titleSmall => typography.titleSmall;
+  String get labelLargeFamily => typography.labelLargeFamily;
+  TextStyle get labelLarge => typography.labelLarge;
+  String get labelMediumFamily => typography.labelMediumFamily;
+  TextStyle get labelMedium => typography.labelMedium;
+  String get labelSmallFamily => typography.labelSmallFamily;
+  TextStyle get labelSmall => typography.labelSmall;
+  String get bodyLargeFamily => typography.bodyLargeFamily;
+  TextStyle get bodyLarge => typography.bodyLarge;
+  String get bodyMediumFamily => typography.bodyMediumFamily;
+  TextStyle get bodyMedium => typography.bodyMedium;
+  String get bodySmallFamily => typography.bodySmallFamily;
+  TextStyle get bodySmall => typography.bodySmall;
+
+  Typography get typography => ThemeTypography(this);
 }
 
 class LightModeTheme extends FlutterFlowTheme {
-  Color primaryColor = const Color(0xFFE8602F);
-  Color secondaryColor = const Color(0xFF000000);
-  Color tertiaryColor = const Color(0xFF333333);
-  Color alternate = const Color(0xFFFF5963);
-  Color primaryBackground = const Color(0xFFF1F4F8);
-  Color secondaryBackground = const Color(0xFFFFFFFF);
-  Color primaryText = const Color(0xFF101213);
-  Color secondaryText = const Color(0xFF57636C);
+  @Deprecated('Use primary instead')
+  Color get primaryColor => primary;
+  @Deprecated('Use secondary instead')
+  Color get secondaryColor => secondary;
+  @Deprecated('Use tertiary instead')
+  Color get tertiaryColor => tertiary;
 
-  Color primaryBtnText = Color(0xFFFFFFFF);
-  Color lineColor = Color(0xFFE0E3E7);
-  Color grayIcon = Color(0xFF95A1AC);
-  Color gray200 = Color(0xFFDBE2E7);
-  Color gray600 = Color(0xFF262D34);
-  Color black600 = Color(0xFF090F13);
-  Color tertiary400 = Color(0xFF39D2C0);
-  Color textColor = Color(0xFF1E2429);
-  Color backgroundComponents = Color(0xFF1D2428);
-  Color headerGray = Color(0xFFE7E7E7);
+  late Color primary = const Color(0xFFE8602F);
+  late Color secondary = const Color(0xFF000000);
+  late Color tertiary = const Color(0xFF333333);
+  late Color alternate = const Color(0xFFFF5963);
+  late Color primaryText = const Color(0xFF101213);
+  late Color secondaryText = const Color(0xFFFFFFFF);
+  late Color primaryBackground = const Color(0xFFF1F4F8);
+  late Color secondaryBackground = const Color(0xFFFFFFFF);
+  late Color accent1 = const Color(0xFF616161);
+  late Color accent2 = const Color(0xFF757575);
+  late Color accent3 = const Color(0xFFE0E0E0);
+  late Color accent4 = const Color(0xFFEEEEEE);
+  late Color success = const Color(0xFF04A24C);
+  late Color warning = const Color(0xFFFCDC0C);
+  late Color error = const Color(0xFFE21C3D);
+  late Color info = const Color(0xFF1C4494);
+
+  late Color primaryBtnText = Color(0xFFFFFFFF);
+  late Color lineColor = Color(0xFFE0E3E7);
+  late Color grayIcon = Color(0xFF95A1AC);
+  late Color gray200 = Color(0xFFDBE2E7);
+  late Color gray600 = Color(0xFF262D34);
+  late Color black600 = Color(0xFF090F13);
+  late Color tertiary400 = Color(0xFF39D2C0);
+  late Color textColor = Color(0xFF1E2429);
+  late Color backgroundComponents = Color(0xFF1D2428);
+  late Color headerGray = Color(0xFFE7E7E7);
+  late Color btnText = Color(0xFFFFFFFF);
+  late Color customColor3 = Color(0xFFDF3F3F);
+  late Color customColor4 = Color(0xFF090F13);
+  late Color white = Color(0xFFFFFFFF);
+  late Color background = Color(0xFF1D2429);
+  late Color customColor1 = Color(0xFF2FB73C);
+  late Color overlay = Color(0xB3FFFFFF);
+  late Color primary600 = Color(0xFF336A4A);
+  late Color secondary600 = Color(0xFF6D604A);
+  late Color tertiary600 = Color(0xFF0C2533);
+  late Color darkBGstatic = Color(0xFF0D1E23);
+  late Color secondary30 = Color(0x4D928163);
+  late Color overlay0 = Color(0x00FFFFFF);
+  late Color primary30 = Color(0x4D4B986C);
+}
+
+abstract class Typography {
+  String get displayLargeFamily;
+  TextStyle get displayLarge;
+  String get displayMediumFamily;
+  TextStyle get displayMedium;
+  String get displaySmallFamily;
+  TextStyle get displaySmall;
+  String get headlineLargeFamily;
+  TextStyle get headlineLarge;
+  String get headlineMediumFamily;
+  TextStyle get headlineMedium;
+  String get headlineSmallFamily;
+  TextStyle get headlineSmall;
+  String get titleLargeFamily;
+  TextStyle get titleLarge;
+  String get titleMediumFamily;
+  TextStyle get titleMedium;
+  String get titleSmallFamily;
+  TextStyle get titleSmall;
+  String get labelLargeFamily;
+  TextStyle get labelLarge;
+  String get labelMediumFamily;
+  TextStyle get labelMedium;
+  String get labelSmallFamily;
+  TextStyle get labelSmall;
+  String get bodyLargeFamily;
+  TextStyle get bodyLarge;
+  String get bodyMediumFamily;
+  TextStyle get bodyMedium;
+  String get bodySmallFamily;
+  TextStyle get bodySmall;
+}
+
+class ThemeTypography extends Typography {
+  ThemeTypography(this.theme);
+
+  final FlutterFlowTheme theme;
+
+  String get displayLargeFamily => 'Lato';
+  TextStyle get displayLarge => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.normal,
+        fontSize: 57.0,
+      );
+  String get displayMediumFamily => 'Lato';
+  TextStyle get displayMedium => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.normal,
+        fontSize: 45.0,
+      );
+  String get displaySmallFamily => 'Lato';
+  TextStyle get displaySmall => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.bold,
+        fontSize: 24.0,
+        fontStyle: FontStyle.normal,
+      );
+  String get headlineLargeFamily => 'Lato';
+  TextStyle get headlineLarge => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.normal,
+        fontSize: 32.0,
+      );
+  String get headlineMediumFamily => 'Lato';
+  TextStyle get headlineMedium => GoogleFonts.getFont(
+        'Lato',
+        color: theme.secondaryText,
+        fontWeight: FontWeight.bold,
+        fontSize: 24.0,
+      );
+  String get headlineSmallFamily => 'Lato';
+  TextStyle get headlineSmall => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 20.0,
+      );
+  String get titleLargeFamily => 'Lato';
+  TextStyle get titleLarge => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w500,
+        fontSize: 22.0,
+      );
+  String get titleMediumFamily => 'Lato';
+  TextStyle get titleMedium => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 18.0,
+      );
+  String get titleSmallFamily => 'Lato';
+  TextStyle get titleSmall => GoogleFonts.getFont(
+        'Lato',
+        color: theme.secondaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 18.0,
+      );
+  String get labelLargeFamily => 'Lato';
+  TextStyle get labelLarge => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w500,
+        fontSize: 14.0,
+      );
+  String get labelMediumFamily => 'Lato';
+  TextStyle get labelMedium => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w500,
+        fontSize: 12.0,
+      );
+  String get labelSmallFamily => 'Lato';
+  TextStyle get labelSmall => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w500,
+        fontSize: 11.0,
+      );
+  String get bodyLargeFamily => '';
+  TextStyle get bodyLarge => GoogleFonts.getFont(
+        'Roboto',
+      );
+  String get bodyMediumFamily => 'Lato';
+  TextStyle get bodyMedium => GoogleFonts.getFont(
+        'Lato',
+        color: theme.primaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 14.0,
+      );
+  String get bodySmallFamily => 'Lato';
+  TextStyle get bodySmall => GoogleFonts.getFont(
+        'Lato',
+        color: theme.secondaryText,
+        fontWeight: FontWeight.w600,
+        fontSize: 14.0,
+      );
 }
 
 class DarkModeTheme extends FlutterFlowTheme {
-  Color primaryColor = const Color(0xFFE8602F);
-  Color secondaryColor = const Color(0xFF000000);
-  Color tertiaryColor = const Color(0xFF333333);
-  Color alternate = const Color(0xFFFF5963);
-  Color primaryBackground = const Color(0xFF1A1F24);
-  Color secondaryBackground = const Color(0xFF101213);
-  Color primaryText = const Color(0xFFFFFFFF);
-  Color secondaryText = const Color(0xFF95A1AC);
+  @Deprecated('Use primary instead')
+  Color get primaryColor => primary;
+  @Deprecated('Use secondary instead')
+  Color get secondaryColor => secondary;
+  @Deprecated('Use tertiary instead')
+  Color get tertiaryColor => tertiary;
 
-  Color primaryBtnText = Color(0xFFFFFFFF);
-  Color lineColor = Color(0xFF22282F);
-  Color grayIcon = Color(0xFF95A1AC);
-  Color gray200 = Color(0xFFDBE2E7);
-  Color gray600 = Color(0xFF262D34);
-  Color black600 = Color(0xFF090F13);
-  Color tertiary400 = Color(0xFF39D2C0);
-  Color textColor = Color(0xFF1E2429);
-  Color backgroundComponents = Color(0xFF1D2428);
-  Color headerGray = Color(0x03E3E3E3);
+  late Color primary = const Color(0xFFE8602F);
+  late Color secondary = const Color(0xFF000000);
+  late Color tertiary = const Color(0xFF333333);
+  late Color alternate = const Color(0xFFFF5963);
+  late Color primaryText = const Color(0xFFFFFFFF);
+  late Color secondaryText = const Color(0xFFFFFFFF);
+  late Color primaryBackground = const Color(0xFF000000);
+  late Color secondaryBackground = const Color(0xFF101213);
+  late Color accent1 = const Color(0xFFEEEEEE);
+  late Color accent2 = const Color(0xFFE0E0E0);
+  late Color accent3 = const Color(0xFF757575);
+  late Color accent4 = const Color(0xFF616161);
+  late Color success = const Color(0xFF04A24C);
+  late Color warning = const Color(0xFFFCDC0C);
+  late Color error = const Color(0xFFE21C3D);
+  late Color info = const Color(0xFF1C4494);
+
+  late Color primaryBtnText = Color(0xFFFFFFFF);
+  late Color lineColor = Color(0xFF22282F);
+  late Color grayIcon = Color(0xFF95A1AC);
+  late Color gray200 = Color(0xFFDBE2E7);
+  late Color gray600 = Color(0xFF262D34);
+  late Color black600 = Color(0xFF090F13);
+  late Color tertiary400 = Color(0xFF39D2C0);
+  late Color textColor = Color(0xFF1E2429);
+  late Color backgroundComponents = Color(0xFF1D2428);
+  late Color headerGray = Color(0x03E3E3E3);
+  late Color btnText = Color(0xFFFFFFFF);
+  late Color customColor3 = Color(0xFFDF3F3F);
+  late Color customColor4 = Color(0xFF090F13);
+  late Color white = Color(0xFFFFFFFF);
+  late Color background = Color(0xFF1D2429);
+  late Color customColor1 = Color(0xFF452FB7);
+  late Color overlay = Color(0xB314181B);
+  late Color primary600 = Color(0xFF336A4A);
+  late Color secondary600 = Color(0xFF6D604A);
+  late Color tertiary600 = Color(0xFF0C2533);
+  late Color darkBGstatic = Color(0xFF0D1E23);
+  late Color secondary30 = Color(0x4D928163);
+  late Color overlay0 = Color(0x000B191E);
+  late Color primary30 = Color(0x4D4B986C);
 }
 
 extension TextStyleHelper on TextStyle {
   TextStyle override({
-    String fontFamily,
-    Color color,
-    double fontSize,
-    FontWeight fontWeight,
-    FontStyle fontStyle,
+    String? fontFamily,
+    Color? color,
+    double? fontSize,
+    FontWeight? fontWeight,
+    double? letterSpacing,
+    FontStyle? fontStyle,
     bool useGoogleFonts = true,
-    TextDecoration decoration,
-    double lineHeight,
+    TextDecoration? decoration,
+    double? lineHeight,
   }) =>
       useGoogleFonts
           ? GoogleFonts.getFont(
-              fontFamily,
+              fontFamily!,
               color: color ?? this.color,
               fontSize: fontSize ?? this.fontSize,
+              letterSpacing: letterSpacing ?? this.letterSpacing,
               fontWeight: fontWeight ?? this.fontWeight,
               fontStyle: fontStyle ?? this.fontStyle,
               decoration: decoration,
@@ -162,6 +414,7 @@ extension TextStyleHelper on TextStyle {
               fontFamily: fontFamily,
               color: color,
               fontSize: fontSize,
+              letterSpacing: letterSpacing,
               fontWeight: fontWeight,
               fontStyle: fontStyle,
               decoration: decoration,
