@@ -1,26 +1,44 @@
-import 'package:firebase_core/firebase_core.dart';
+import 'package:provider/provider.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'auth/firebase_user_provider.dart';
-import 'auth/auth_util.dart';
+import 'package:flutter_web_plugins/url_strategy.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'auth/firebase_auth/firebase_user_provider.dart';
+import 'auth/firebase_auth/auth_util.dart';
+
 import 'backend/push_notifications/push_notifications_util.dart';
+import 'backend/firebase/firebase_config.dart';
 import 'flutter_flow/flutter_flow_theme.dart';
 import 'flutter_flow/flutter_flow_util.dart';
 import 'flutter_flow/internationalization.dart';
-import 'package:flutter_spinkit/flutter_spinkit.dart';
+import 'package:flutter/foundation.dart';
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:google_nav_bar/google_nav_bar.dart';
 import 'flutter_flow/nav/nav.dart';
 import 'index.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp();
+  usePathUrlStrategy();
+  await initFirebase();
+
   await FlutterFlowTheme.initialize();
 
-  FFAppState(); // Initialize FFAppState
+  final appState = FFAppState(); // Initialize FFAppState
+  await appState.initializePersistedState();
 
-  runApp(MyApp());
+  if (!kIsWeb) {
+    FlutterError.onError = FirebaseCrashlytics.instance.recordFlutterFatalError;
+  }
+  await initializeFirebaseRemoteConfig();
+
+  runApp(ChangeNotifierProvider(
+    create: (context) => appState,
+    child: MyApp(),
+  ));
 }
 
 class MyApp extends StatefulWidget {
@@ -29,17 +47,17 @@ class MyApp extends StatefulWidget {
   State<MyApp> createState() => _MyAppState();
 
   static _MyAppState of(BuildContext context) =>
-      context.findAncestorStateOfType<_MyAppState>();
+      context.findAncestorStateOfType<_MyAppState>()!;
 }
 
 class _MyAppState extends State<MyApp> {
-  Locale _locale;
+  Locale? _locale;
   ThemeMode _themeMode = FlutterFlowTheme.themeMode;
 
-  Stream<LiveTVFirebaseUser> userStream;
+  late Stream<BaseAuthUser> userStream;
 
-  AppStateNotifier _appStateNotifier;
-  GoRouter _router;
+  late AppStateNotifier _appStateNotifier;
+  late GoRouter _router;
 
   final authUserSub = authenticatedUserStream.listen((_) {});
   final fcmTokenSub = fcmTokenUserStream.listen((_) {});
@@ -47,10 +65,11 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     super.initState();
-    _appStateNotifier = AppStateNotifier();
+    _appStateNotifier = AppStateNotifier.instance;
     _router = createRouter(_appStateNotifier);
     userStream = liveTVFirebaseUserStream()
       ..listen((user) => _appStateNotifier.update(user));
+    jwtTokenStream.listen((_) {});
     Future.delayed(
       Duration(seconds: 1),
       () => _appStateNotifier.stopShowingSplashImage(),
@@ -64,7 +83,10 @@ class _MyAppState extends State<MyApp> {
     super.dispose();
   }
 
-  void setLocale(Locale value) => setState(() => _locale = value);
+  void setLocale(String language) {
+    setState(() => _locale = createLocale(language));
+  }
+
   void setThemeMode(ThemeMode mode) => setState(() {
         _themeMode = mode;
         FlutterFlowTheme.saveThemeMode(mode);
@@ -81,20 +103,23 @@ class _MyAppState extends State<MyApp> {
         GlobalCupertinoLocalizations.delegate,
       ],
       locale: _locale,
-      supportedLocales: const [Locale('en', '')],
+      supportedLocales: const [
+        Locale('en'),
+        Locale('es'),
+      ],
       theme: ThemeData(brightness: Brightness.light),
       darkTheme: ThemeData(brightness: Brightness.dark),
       themeMode: _themeMode,
-      routeInformationParser: _router.routeInformationParser,
-      routerDelegate: _router.routerDelegate,
+      routerConfig: _router,
     );
   }
 }
 
 class NavBarPage extends StatefulWidget {
-  NavBarPage({Key key, this.initialPage}) : super(key: key);
+  NavBarPage({Key? key, this.initialPage, this.page}) : super(key: key);
 
-  final String initialPage;
+  final String? initialPage;
+  final Widget? page;
 
   @override
   _NavBarPageState createState() => _NavBarPageState();
@@ -102,92 +127,89 @@ class NavBarPage extends StatefulWidget {
 
 /// This is the private State class that goes with NavBarPage.
 class _NavBarPageState extends State<NavBarPage> {
-  String _currentPage = 'Home';
+  String _currentPageName = 'Home';
+  late Widget? _currentPage;
 
   @override
   void initState() {
     super.initState();
-    _currentPage = widget.initialPage ?? _currentPage;
+    _currentPageName = widget.initialPage ?? _currentPageName;
+    _currentPage = widget.page;
   }
 
   @override
   Widget build(BuildContext context) {
     final tabs = {
       'Home': HomeWidget(),
-      'DailyDevoCal': DailyDevoCalWidget(),
-      'SermonNav': SermonNavWidget(),
       'Bible': BibleWidget(),
+      'DailyDevoCal': DailyDevoCalWidget(),
       'allChats': AllChatsWidget(),
       'socialFeed': SocialFeedWidget(),
     };
-    final currentIndex = tabs.keys.toList().indexOf(_currentPage);
+    final currentIndex = tabs.keys.toList().indexOf(_currentPageName);
+
     return Scaffold(
-      body: tabs[_currentPage],
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: currentIndex,
-        onTap: (i) => setState(() => _currentPage = tabs.keys.toList()[i]),
-        backgroundColor: Colors.white,
-        selectedItemColor: Color(0xFFE8602F),
-        unselectedItemColor: Colors.black,
-        showSelectedLabels: true,
-        showUnselectedLabels: true,
-        type: BottomNavigationBarType.fixed,
-        items: <BottomNavigationBarItem>[
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.home_outlined,
-              size: 24,
+      body: _currentPage ?? tabs[_currentPageName],
+      bottomNavigationBar: GNav(
+        selectedIndex: currentIndex,
+        onTabChange: (i) => setState(() {
+          _currentPage = null;
+          _currentPageName = tabs.keys.toList()[i];
+        }),
+        backgroundColor: FlutterFlowTheme.of(context).secondary,
+        color: Colors.white,
+        activeColor: Color(0xFFE8602F),
+        tabBackgroundColor: Colors.white,
+        tabActiveBorder: Border.all(
+          color: FlutterFlowTheme.of(context).primary,
+          width: 3.0,
+        ),
+        tabBorderRadius: 4.0,
+        tabMargin: EdgeInsetsDirectional.fromSTEB(0.0, 5.0, 0.0, 25.0),
+        padding: EdgeInsetsDirectional.fromSTEB(10.0, 10.0, 10.0, 10.0),
+        gap: 5.0,
+        mainAxisAlignment: MainAxisAlignment.spaceAround,
+        duration: Duration(milliseconds: 500),
+        haptic: true,
+        tabs: [
+          GButton(
+            icon: Icons.home_outlined,
+            text: FFLocalizations.of(context).getText(
+              'qqbfqfvt' /* Home */,
             ),
-            label: 'Home',
-            tooltip: '',
+            iconSize: 24.0,
           ),
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.featured_play_list,
-              size: 24,
+          GButton(
+            icon: Icons.menu_book,
+            text: FFLocalizations.of(context).getText(
+              '6bjrduxd' /* Bible */,
             ),
-            label: 'Daily',
-            tooltip: '',
+            iconSize: 24.0,
           ),
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.personal_video,
-              size: 24,
+          GButton(
+            icon: currentIndex == 2
+                ? Icons.featured_play_list_outlined
+                : Icons.featured_play_list,
+            text: FFLocalizations.of(context).getText(
+              'zwihpbx0' /* Daily */,
             ),
-            activeIcon: Icon(
-              Icons.live_tv_sharp,
-              size: 24,
-            ),
-            label: 'Sermons',
-            tooltip: '',
+            iconSize: 24.0,
           ),
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.menu_book,
-              size: 24,
+          GButton(
+            icon: currentIndex == 3
+                ? Icons.chat_bubble_rounded
+                : Icons.chat_bubble_outline,
+            text: FFLocalizations.of(context).getText(
+              'bebawnx3' /* Chats */,
             ),
-            label: 'Bible',
-            tooltip: '',
+            iconSize: 24.0,
           ),
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.chat_bubble_outline,
-              size: 24,
+          GButton(
+            icon: Icons.person,
+            text: FFLocalizations.of(context).getText(
+              'xuxmy966' /* Social */,
             ),
-            activeIcon: Icon(
-              Icons.chat_bubble_rounded,
-              size: 24,
-            ),
-            label: 'Chats',
-            tooltip: '',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.person,
-              size: 24,
-            ),
-            label: 'Social',
-            tooltip: '',
+            iconSize: 24.0,
           )
         ],
       ),
