@@ -1,7 +1,9 @@
 import 'package:cloud_functions/cloud_functions.dart';
 
 Future<Map<String, dynamic>> makeCloudCall(
-    String callName, Map<String, dynamic> input) async {
+  String callName,
+  Map<String, dynamic> input,
+) async {
   try {
     final response = await FirebaseFunctions.instance
         .httpsCallable(callName, options: HttpsCallableOptions())
