@@ -1,73 +1,125 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart';
+
+import '/backend/schema/util/firestore_util.dart';
+import '/backend/schema/util/schema_util.dart';
+
 import 'index.dart';
-import 'serializers.dart';
-import 'package:built_value/built_value.dart';
+import '/flutter_flow/flutter_flow_util.dart';
 
-part 'chat_messages_record.g.dart';
+class ChatMessagesRecord extends FirestoreRecord {
+  ChatMessagesRecord._(
+    DocumentReference reference,
+    Map<String, dynamic> data,
+  ) : super(reference, data) {
+    _initializeFields();
+  }
 
-abstract class ChatMessagesRecord
-    implements Built<ChatMessagesRecord, ChatMessagesRecordBuilder> {
-  static Serializer<ChatMessagesRecord> get serializer =>
-      _$chatMessagesRecordSerializer;
+  // "user" field.
+  DocumentReference? _user;
+  DocumentReference? get user => _user;
+  bool hasUser() => _user != null;
 
-  @nullable
-  DocumentReference get user;
+  // "chat" field.
+  DocumentReference? _chat;
+  DocumentReference? get chat => _chat;
+  bool hasChat() => _chat != null;
 
-  @nullable
-  DocumentReference get chat;
+  // "text" field.
+  String? _text;
+  String get text => _text ?? '';
+  bool hasText() => _text != null;
 
-  @nullable
-  String get text;
+  // "image" field.
+  String? _image;
+  String get image => _image ?? '';
+  bool hasImage() => _image != null;
 
-  @nullable
-  String get image;
+  // "timestamp" field.
+  DateTime? _timestamp;
+  DateTime? get timestamp => _timestamp;
+  bool hasTimestamp() => _timestamp != null;
 
-  @nullable
-  DateTime get timestamp;
-
-  @nullable
-  @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
-
-  static void _initializeBuilder(ChatMessagesRecordBuilder builder) => builder
-    ..text = ''
-    ..image = '';
+  void _initializeFields() {
+    _user = snapshotData['user'] as DocumentReference?;
+    _chat = snapshotData['chat'] as DocumentReference?;
+    _text = snapshotData['text'] as String?;
+    _image = snapshotData['image'] as String?;
+    _timestamp = snapshotData['timestamp'] as DateTime?;
+  }
 
   static CollectionReference get collection =>
       FirebaseFirestore.instance.collection('chat_messages');
 
-  static Stream<ChatMessagesRecord> getDocument(DocumentReference ref) => ref
-      .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+  static Stream<ChatMessagesRecord> getDocument(DocumentReference ref) =>
+      ref.snapshots().map((s) => ChatMessagesRecord.fromSnapshot(s));
 
   static Future<ChatMessagesRecord> getDocumentOnce(DocumentReference ref) =>
-      ref.get().then(
-          (s) => serializers.deserializeWith(serializer, serializedData(s)));
+      ref.get().then((s) => ChatMessagesRecord.fromSnapshot(s));
 
-  ChatMessagesRecord._();
-  factory ChatMessagesRecord(
-          [void Function(ChatMessagesRecordBuilder) updates]) =
-      _$ChatMessagesRecord;
+  static ChatMessagesRecord fromSnapshot(DocumentSnapshot snapshot) =>
+      ChatMessagesRecord._(
+        snapshot.reference,
+        mapFromFirestore(snapshot.data() as Map<String, dynamic>),
+      );
 
   static ChatMessagesRecord getDocumentFromData(
-          Map<String, dynamic> data, DocumentReference reference) =>
-      serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+    Map<String, dynamic> data,
+    DocumentReference reference,
+  ) =>
+      ChatMessagesRecord._(reference, mapFromFirestore(data));
+
+  @override
+  String toString() =>
+      'ChatMessagesRecord(reference: ${reference.path}, data: $snapshotData)';
+
+  @override
+  int get hashCode => reference.path.hashCode;
+
+  @override
+  bool operator ==(other) =>
+      other is ChatMessagesRecord &&
+      reference.path.hashCode == other.reference.path.hashCode;
 }
 
 Map<String, dynamic> createChatMessagesRecordData({
-  DocumentReference user,
-  DocumentReference chat,
-  String text,
-  String image,
-  DateTime timestamp,
-}) =>
-    serializers.toFirestore(
-        ChatMessagesRecord.serializer,
-        ChatMessagesRecord((c) => c
-          ..user = user
-          ..chat = chat
-          ..text = text
-          ..image = image
-          ..timestamp = timestamp));
+  DocumentReference? user,
+  DocumentReference? chat,
+  String? text,
+  String? image,
+  DateTime? timestamp,
+}) {
+  final firestoreData = mapToFirestore(
+    <String, dynamic>{
+      'user': user,
+      'chat': chat,
+      'text': text,
+      'image': image,
+      'timestamp': timestamp,
+    }.withoutNulls,
+  );
+
+  return firestoreData;
+}
+
+class ChatMessagesRecordDocumentEquality
+    implements Equality<ChatMessagesRecord> {
+  const ChatMessagesRecordDocumentEquality();
+
+  @override
+  bool equals(ChatMessagesRecord? e1, ChatMessagesRecord? e2) {
+    return e1?.user == e2?.user &&
+        e1?.chat == e2?.chat &&
+        e1?.text == e2?.text &&
+        e1?.image == e2?.image &&
+        e1?.timestamp == e2?.timestamp;
+  }
+
+  @override
+  int hash(ChatMessagesRecord? e) => const ListEquality()
+      .hash([e?.user, e?.chat, e?.text, e?.image, e?.timestamp]);
+
+  @override
+  bool isValidKey(Object? o) => o is ChatMessagesRecord;
+}
