@@ -1,85 +1,147 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart';
+
+import '/backend/schema/util/firestore_util.dart';
+import '/backend/schema/util/schema_util.dart';
+
 import 'index.dart';
-import 'serializers.dart';
-import 'package:built_value/built_value.dart';
+import '/flutter_flow/flutter_flow_util.dart';
 
-part 'chats_record.g.dart';
+class ChatsRecord extends FirestoreRecord {
+  ChatsRecord._(
+    DocumentReference reference,
+    Map<String, dynamic> data,
+  ) : super(reference, data) {
+    _initializeFields();
+  }
 
-abstract class ChatsRecord implements Built<ChatsRecord, ChatsRecordBuilder> {
-  static Serializer<ChatsRecord> get serializer => _$chatsRecordSerializer;
+  // "users" field.
+  List<DocumentReference>? _users;
+  List<DocumentReference> get users => _users ?? const [];
+  bool hasUsers() => _users != null;
 
-  @nullable
-  BuiltList<DocumentReference> get users;
+  // "user_a" field.
+  DocumentReference? _userA;
+  DocumentReference? get userA => _userA;
+  bool hasUserA() => _userA != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'user_a')
-  DocumentReference get userA;
+  // "user_b" field.
+  DocumentReference? _userB;
+  DocumentReference? get userB => _userB;
+  bool hasUserB() => _userB != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'user_b')
-  DocumentReference get userB;
+  // "last_message" field.
+  String? _lastMessage;
+  String get lastMessage => _lastMessage ?? '';
+  bool hasLastMessage() => _lastMessage != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'last_message')
-  String get lastMessage;
+  // "last_message_time" field.
+  DateTime? _lastMessageTime;
+  DateTime? get lastMessageTime => _lastMessageTime;
+  bool hasLastMessageTime() => _lastMessageTime != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'last_message_time')
-  DateTime get lastMessageTime;
+  // "last_message_sent_by" field.
+  DocumentReference? _lastMessageSentBy;
+  DocumentReference? get lastMessageSentBy => _lastMessageSentBy;
+  bool hasLastMessageSentBy() => _lastMessageSentBy != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'last_message_sent_by')
-  DocumentReference get lastMessageSentBy;
+  // "last_message_seen_by" field.
+  List<DocumentReference>? _lastMessageSeenBy;
+  List<DocumentReference> get lastMessageSeenBy =>
+      _lastMessageSeenBy ?? const [];
+  bool hasLastMessageSeenBy() => _lastMessageSeenBy != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'last_message_seen_by')
-  BuiltList<DocumentReference> get lastMessageSeenBy;
-
-  @nullable
-  @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
-
-  static void _initializeBuilder(ChatsRecordBuilder builder) => builder
-    ..users = ListBuilder()
-    ..lastMessage = ''
-    ..lastMessageSeenBy = ListBuilder();
+  void _initializeFields() {
+    _users = getDataList(snapshotData['users']);
+    _userA = snapshotData['user_a'] as DocumentReference?;
+    _userB = snapshotData['user_b'] as DocumentReference?;
+    _lastMessage = snapshotData['last_message'] as String?;
+    _lastMessageTime = snapshotData['last_message_time'] as DateTime?;
+    _lastMessageSentBy =
+        snapshotData['last_message_sent_by'] as DocumentReference?;
+    _lastMessageSeenBy = getDataList(snapshotData['last_message_seen_by']);
+  }
 
   static CollectionReference get collection =>
       FirebaseFirestore.instance.collection('chats');
 
-  static Stream<ChatsRecord> getDocument(DocumentReference ref) => ref
-      .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+  static Stream<ChatsRecord> getDocument(DocumentReference ref) =>
+      ref.snapshots().map((s) => ChatsRecord.fromSnapshot(s));
 
-  static Future<ChatsRecord> getDocumentOnce(DocumentReference ref) => ref
-      .get()
-      .then((s) => serializers.deserializeWith(serializer, serializedData(s)));
+  static Future<ChatsRecord> getDocumentOnce(DocumentReference ref) =>
+      ref.get().then((s) => ChatsRecord.fromSnapshot(s));
 
-  ChatsRecord._();
-  factory ChatsRecord([void Function(ChatsRecordBuilder) updates]) =
-      _$ChatsRecord;
+  static ChatsRecord fromSnapshot(DocumentSnapshot snapshot) => ChatsRecord._(
+        snapshot.reference,
+        mapFromFirestore(snapshot.data() as Map<String, dynamic>),
+      );
 
   static ChatsRecord getDocumentFromData(
-          Map<String, dynamic> data, DocumentReference reference) =>
-      serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+    Map<String, dynamic> data,
+    DocumentReference reference,
+  ) =>
+      ChatsRecord._(reference, mapFromFirestore(data));
+
+  @override
+  String toString() =>
+      'ChatsRecord(reference: ${reference.path}, data: $snapshotData)';
+
+  @override
+  int get hashCode => reference.path.hashCode;
+
+  @override
+  bool operator ==(other) =>
+      other is ChatsRecord &&
+      reference.path.hashCode == other.reference.path.hashCode;
 }
 
 Map<String, dynamic> createChatsRecordData({
-  DocumentReference userA,
-  DocumentReference userB,
-  String lastMessage,
-  DateTime lastMessageTime,
-  DocumentReference lastMessageSentBy,
-}) =>
-    serializers.toFirestore(
-        ChatsRecord.serializer,
-        ChatsRecord((c) => c
-          ..users = null
-          ..userA = userA
-          ..userB = userB
-          ..lastMessage = lastMessage
-          ..lastMessageTime = lastMessageTime
-          ..lastMessageSentBy = lastMessageSentBy
-          ..lastMessageSeenBy = null));
+  DocumentReference? userA,
+  DocumentReference? userB,
+  String? lastMessage,
+  DateTime? lastMessageTime,
+  DocumentReference? lastMessageSentBy,
+}) {
+  final firestoreData = mapToFirestore(
+    <String, dynamic>{
+      'user_a': userA,
+      'user_b': userB,
+      'last_message': lastMessage,
+      'last_message_time': lastMessageTime,
+      'last_message_sent_by': lastMessageSentBy,
+    }.withoutNulls,
+  );
+
+  return firestoreData;
+}
+
+class ChatsRecordDocumentEquality implements Equality<ChatsRecord> {
+  const ChatsRecordDocumentEquality();
+
+  @override
+  bool equals(ChatsRecord? e1, ChatsRecord? e2) {
+    const listEquality = ListEquality();
+    return listEquality.equals(e1?.users, e2?.users) &&
+        e1?.userA == e2?.userA &&
+        e1?.userB == e2?.userB &&
+        e1?.lastMessage == e2?.lastMessage &&
+        e1?.lastMessageTime == e2?.lastMessageTime &&
+        e1?.lastMessageSentBy == e2?.lastMessageSentBy &&
+        listEquality.equals(e1?.lastMessageSeenBy, e2?.lastMessageSeenBy);
+  }
+
+  @override
+  int hash(ChatsRecord? e) => const ListEquality().hash([
+        e?.users,
+        e?.userA,
+        e?.userB,
+        e?.lastMessage,
+        e?.lastMessageTime,
+        e?.lastMessageSentBy,
+        e?.lastMessageSeenBy
+      ]);
+
+  @override
+  bool isValidKey(Object? o) => o is ChatsRecord;
+}
