@@ -1,95 +1,166 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart';
+
+import '/backend/schema/util/firestore_util.dart';
+import '/backend/schema/util/schema_util.dart';
+
 import 'index.dart';
-import 'serializers.dart';
-import 'package:built_value/built_value.dart';
+import '/flutter_flow/flutter_flow_util.dart';
 
-part 'social_posts_record.g.dart';
+class SocialPostsRecord extends FirestoreRecord {
+  SocialPostsRecord._(
+    DocumentReference reference,
+    Map<String, dynamic> data,
+  ) : super(reference, data) {
+    _initializeFields();
+  }
 
-abstract class SocialPostsRecord
-    implements Built<SocialPostsRecord, SocialPostsRecordBuilder> {
-  static Serializer<SocialPostsRecord> get serializer =>
-      _$socialPostsRecordSerializer;
+  // "content" field.
+  String? _content;
+  String get content => _content ?? '';
+  bool hasContent() => _content != null;
 
-  @nullable
-  String get content;
+  // "likes" field.
+  List<DocumentReference>? _likes;
+  List<DocumentReference> get likes => _likes ?? const [];
+  bool hasLikes() => _likes != null;
 
-  @nullable
-  BuiltList<DocumentReference> get likes;
+  // "photo" field.
+  String? _photo;
+  String get photo => _photo ?? '';
+  bool hasPhoto() => _photo != null;
 
-  @nullable
-  String get photo;
+  // "posted_by" field.
+  DocumentReference? _postedBy;
+  DocumentReference? get postedBy => _postedBy;
+  bool hasPostedBy() => _postedBy != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'posted_by')
-  DocumentReference get postedBy;
+  // "posted_time" field.
+  DateTime? _postedTime;
+  DateTime? get postedTime => _postedTime;
+  bool hasPostedTime() => _postedTime != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'posted_time')
-  DateTime get postedTime;
+  // "video_link" field.
+  String? _videoLink;
+  String get videoLink => _videoLink ?? '';
+  bool hasVideoLink() => _videoLink != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'video_link')
-  String get videoLink;
+  // "video_name" field.
+  String? _videoName;
+  String get videoName => _videoName ?? '';
+  bool hasVideoName() => _videoName != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'video_name')
-  String get videoName;
+  // "is_video_post" field.
+  bool? _isVideoPost;
+  bool get isVideoPost => _isVideoPost ?? false;
+  bool hasIsVideoPost() => _isVideoPost != null;
 
-  @nullable
-  @BuiltValueField(wireName: 'is_video_post')
-  bool get isVideoPost;
+  // "liked_by" field.
+  List<DocumentReference>? _likedBy;
+  List<DocumentReference> get likedBy => _likedBy ?? const [];
+  bool hasLikedBy() => _likedBy != null;
 
-  @nullable
-  @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
-
-  static void _initializeBuilder(SocialPostsRecordBuilder builder) => builder
-    ..content = ''
-    ..likes = ListBuilder()
-    ..photo = ''
-    ..videoLink = ''
-    ..videoName = ''
-    ..isVideoPost = false;
+  void _initializeFields() {
+    _content = snapshotData['content'] as String?;
+    _likes = getDataList(snapshotData['likes']);
+    _photo = snapshotData['photo'] as String?;
+    _postedBy = snapshotData['posted_by'] as DocumentReference?;
+    _postedTime = snapshotData['posted_time'] as DateTime?;
+    _videoLink = snapshotData['video_link'] as String?;
+    _videoName = snapshotData['video_name'] as String?;
+    _isVideoPost = snapshotData['is_video_post'] as bool?;
+    _likedBy = getDataList(snapshotData['liked_by']);
+  }
 
   static CollectionReference get collection =>
       FirebaseFirestore.instance.collection('social_posts');
 
-  static Stream<SocialPostsRecord> getDocument(DocumentReference ref) => ref
-      .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+  static Stream<SocialPostsRecord> getDocument(DocumentReference ref) =>
+      ref.snapshots().map((s) => SocialPostsRecord.fromSnapshot(s));
 
-  static Future<SocialPostsRecord> getDocumentOnce(DocumentReference ref) => ref
-      .get()
-      .then((s) => serializers.deserializeWith(serializer, serializedData(s)));
+  static Future<SocialPostsRecord> getDocumentOnce(DocumentReference ref) =>
+      ref.get().then((s) => SocialPostsRecord.fromSnapshot(s));
 
-  SocialPostsRecord._();
-  factory SocialPostsRecord([void Function(SocialPostsRecordBuilder) updates]) =
-      _$SocialPostsRecord;
+  static SocialPostsRecord fromSnapshot(DocumentSnapshot snapshot) =>
+      SocialPostsRecord._(
+        snapshot.reference,
+        mapFromFirestore(snapshot.data() as Map<String, dynamic>),
+      );
 
   static SocialPostsRecord getDocumentFromData(
-          Map<String, dynamic> data, DocumentReference reference) =>
-      serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+    Map<String, dynamic> data,
+    DocumentReference reference,
+  ) =>
+      SocialPostsRecord._(reference, mapFromFirestore(data));
+
+  @override
+  String toString() =>
+      'SocialPostsRecord(reference: ${reference.path}, data: $snapshotData)';
+
+  @override
+  int get hashCode => reference.path.hashCode;
+
+  @override
+  bool operator ==(other) =>
+      other is SocialPostsRecord &&
+      reference.path.hashCode == other.reference.path.hashCode;
 }
 
 Map<String, dynamic> createSocialPostsRecordData({
-  String content,
-  String photo,
-  DocumentReference postedBy,
-  DateTime postedTime,
-  String videoLink,
-  String videoName,
-  bool isVideoPost,
-}) =>
-    serializers.toFirestore(
-        SocialPostsRecord.serializer,
-        SocialPostsRecord((s) => s
-          ..content = content
-          ..likes = null
-          ..photo = photo
-          ..postedBy = postedBy
-          ..postedTime = postedTime
-          ..videoLink = videoLink
-          ..videoName = videoName
-          ..isVideoPost = isVideoPost));
+  String? content,
+  String? photo,
+  DocumentReference? postedBy,
+  DateTime? postedTime,
+  String? videoLink,
+  String? videoName,
+  bool? isVideoPost,
+}) {
+  final firestoreData = mapToFirestore(
+    <String, dynamic>{
+      'content': content,
+      'photo': photo,
+      'posted_by': postedBy,
+      'posted_time': postedTime,
+      'video_link': videoLink,
+      'video_name': videoName,
+      'is_video_post': isVideoPost,
+    }.withoutNulls,
+  );
+
+  return firestoreData;
+}
+
+class SocialPostsRecordDocumentEquality implements Equality<SocialPostsRecord> {
+  const SocialPostsRecordDocumentEquality();
+
+  @override
+  bool equals(SocialPostsRecord? e1, SocialPostsRecord? e2) {
+    const listEquality = ListEquality();
+    return e1?.content == e2?.content &&
+        listEquality.equals(e1?.likes, e2?.likes) &&
+        e1?.photo == e2?.photo &&
+        e1?.postedBy == e2?.postedBy &&
+        e1?.postedTime == e2?.postedTime &&
+        e1?.videoLink == e2?.videoLink &&
+        e1?.videoName == e2?.videoName &&
+        e1?.isVideoPost == e2?.isVideoPost &&
+        listEquality.equals(e1?.likedBy, e2?.likedBy);
+  }
+
+  @override
+  int hash(SocialPostsRecord? e) => const ListEquality().hash([
+        e?.content,
+        e?.likes,
+        e?.photo,
+        e?.postedBy,
+        e?.postedTime,
+        e?.videoLink,
+        e?.videoName,
+        e?.isVideoPost,
+        e?.likedBy
+      ]);
+
+  @override
+  bool isValidKey(Object? o) => o is SocialPostsRecord;
+}
