@@ -1,8 +1,8 @@
-export '../../auth/auth_util.dart';
-export '../../backend/backend.dart';
-export '../../backend/firebase_storage/storage.dart';
+export '/auth/firebase_auth/auth_util.dart';
+export '/backend/backend.dart';
+export '/backend/firebase_storage/storage.dart';
 export '../flutter_flow_util.dart';
-export '../upload_media.dart';
+export '../upload_data.dart';
 export 'flutter_flow_chat_manager.dart';
 export 'flutter_flow_chat_page.dart';
 export 'flutter_flow_chat_preview.dart';
